pyproject.toml
[tool.poetry]
name = "sevenbot"
version = "0.1.0"
description = "A multifunctional bot for discord"
authors = ["sevenc-nanashi <sevenc-nanashi@sevenbot.jp>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.9"
async-google-trans-new = "~=1.4.2"
aiohttp = "^3.7.4"
authlib = "~=0.15.5"
beautifulsoup4 = "~=4.9.3"
discord-emoji = "~=1.3.1"
discord-ext-levenshtein = "~=0.3.0"
dnspython = "~=2.1.0"
dpy-components = "~=0.4.2"
dpy-peper = "~=1.0.1"
dpy-syntaxer = "~=1.1.0"
google-api-python-client = "~=2.9.0"
httpx = "~=0.18.2"
jishaku = "~=2.0.0"
lxml = "~=4.6.5"
motor = "~=2.4.0"
pillow = "~=9.0.0"
psutil = "~=5.8.0"
py = "~=1.10.0"
pylint = "~=2.8.3"
pynacl = "~=1.4.0"
pyre2 = "~=0.3.6"
"discord.py" = { git = "https://github.com/Rapptz/discord.py.git", rev = "66155faf" }

sembed = "~=1.1.2"
sentry-sdk = "~=1.5.0"
texttable = "~=1.6.4"
topggpy = "~=1.1.0"
websockets = "~=10.1"
xmltodict = "~=0.12.0"
youtube-dl = "~=2021.6.6"

[tool.poetry.dev-dependencies]
pip-chill = "~=1.0.1"
flake8 = "~=4.0.1"
black = "~=21.11b1"
pyproject-flake8 = "^0.0.1-alpha.2"
isort = "^5.10.1"
watchdog = "^2.1.6"

[tool.black]
line-length = 120
extend-exclude = "__gi_*"

[tool.flake8]
max-line-length = 120
extend-ignore = "E203,W503"
extend-exclude = ".venv,__gi_*"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools", "wheel"]
build-backend = "poetry.core.masonry.api"
